Update index.html to display group and piece information
The original static content on the index.html page has been replaced with dynamic content. The new code assumes that it receives a list of 'group' objects, which are then displayed in a selectable dropdown. Each 'group' contains 'piece' objects, which are represented by cards that include an image and description.

Signed-off-by: discovery-bit <61625792+discovery-bit@users.noreply.github.com>